refactor(chart): merge `types.go` into `chart.go` to reduce file fragmentation
